+++ app/build.gradle
@@ -5,11 +5,13 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
+    dataBinding.enabled = true
     defaultConfig {
         applicationId "com.bakkenbaeck.sol"
         minSdkVersion 14
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 6
         versionName "1.1"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -28,8 +30,9 @@ dependencies {
     implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
     // Support
-    implementation "com.android.support:appcompat-v7:26.1.0"
+    implementation 'com.android.support:appcompat-v7:28.0.0'
 
+    // Location awareness
     implementation "com.google.android.gms:play-services-location:16.0.0"
 
     // Prettyness
